fix comparisons for null jsons and geo strings
