+++ flow/connectors/postgres/qvalue_convert.go
@@ -1,7 +1,6 @@
 package connpostgres
 
 import (
-	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -14,8 +13,6 @@ import (
 	"github.com/PeerDB-io/peer-flow/model/qvalue"
 	"github.com/jackc/pgx/v5/pgtype"
 	"github.com/lib/pq/oid"
-
-	geom "github.com/twpayne/go-geos"
 )
 
 func postgresOIDToQValueKind(recvOID uint32) qvalue.QValueKind {
@@ -406,29 +403,4 @@ func customTypeToQKind(typeName string) qvalue.QValueKind {
 		qValueKind = qvalue.QValueKindString
 	}
 	return qValueKind
-}
-
-// returns the WKT representation of the geometry object if it is valid
-func GeoValidate(hexWkb string) (string, error) {
-	// Decode the WKB hex string into binary
-	wkb, hexErr := hex.DecodeString(hexWkb)
-	if hexErr != nil {
-		slog.Warn(fmt.Sprintf("Ignoring invalid WKB: %s", hexWkb))
-		return "", hexErr
-	}
-
-	// UnmarshalWKB performs geometry validation along with WKB parsing
-	geometryObject, geoErr := geom.NewGeomFromWKB(wkb)
-	if geoErr != nil {
-		return "", geoErr
-	}
-
-	invalidReason := geometryObject.IsValidReason()
-	if invalidReason != "Valid Geometry" {
-		slog.Warn(fmt.Sprintf("Ignoring invalid geometry shape %s: %s", hexWkb, invalidReason))
-		return "", errors.New(invalidReason)
-	}
-
-	wkt := geometryObject.ToWKT()
-	return wkt, nil
 }
